Piscine Data Science/1 - Data Warehouse/requirements.txt
pandas==2.3.3
psycopg2-binary==2.9.11